requirements.txt
asgiref==3.7.2
Django==4.2.7
django-filter==23.3
djangogrpcframework==0.2.1
djangorestframework==3.14.0
grpcio==1.59.2
grpcio-tools==1.59.2
Markdown==3.5.1
protobuf==4.25.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
